Update centos8 dockerfile to use local source for container, add entrypoint.sh
Make the centos8 dockerfile to use the local source code rather than
pull the code from github.  This allows us to make changes and test
them in the container.

Make the centos8 dockerfile the same as the ubuntu dockerfile by
adding entrypoint.sh and using it as the entrypoint for the container.