Add debugging for eNB attach failure due to mismatched GUMMEI.
Update s1ap_mme_compare_gummei to add debug message showing the
configured GUMMEI PLMN and the PLMN that is being searched for so
that configuration mismatches can be more easily found.

Use the PLMNS_ARE_EQUAL macro to test for equality for better
readability.

Add missing newline to debug message when matching GUMMEI is not found.